chore - restructured project according to feature
