+++ app/src/main/java/com/judahben149/note/note/adapter/FavoriteNoteListAdapter.kt
@@ -1,12 +1,12 @@
-package com.judahben149.note.adapters
+package com.judahben149.note.note.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.navigation.Navigation
 import androidx.recyclerview.widget.RecyclerView
 import com.judahben149.note.databinding.NoteItemBinding
-import com.judahben149.note.fragments.favoriteNotes.FavoriteNoteListFragmentDirections
-import com.judahben149.note.model.Note
+import com.judahben149.note.note.fragment.addNoteToFavorite.FavoriteNoteListFragmentDirections
+import com.judahben149.note.note.model.Note
 import org.ocpsoft.prettytime.PrettyTime
 import java.util.*
