+++ app/src/main/res/layout/fragment_deleted_note_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.deletedNotes.DeletedNoteListFragment">
+    tools:context=".note.fragment.deletedNote.DeletedNoteListFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
