+++ app/src/main/res/layout/fragment_favorite_note_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.favoriteNotes.FavoriteNoteListFragment">
+    tools:context=".note.fragment.addNoteToFavorite.FavoriteNoteListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvFavoritesNotesList"
